% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{add_samples_tbl}
\alias{add_samples_tbl}
\title{Add Samples Table}
\usage{
add_samples_tbl(mzroll_list, samples_tbl, id_strings, exact = TRUE)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{samples_tbl}{Table of sample metadata}

\item{id_strings}{one or more variables which will be used to match
sample names.}

\item{exact}{if true, an exact match between mzroll names and id_strings
will be found; if false, then substring matches will be used.}
}
\description{
Add a table of sample metadata to an existing mzroll_list
}
\examples{
mzroll_list <- process_mzroll(nplug_mzroll())
add_samples_tbl(mzroll_list, nplug_samples, "sample_name")

}
