% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{aggregate_mzroll_nest}
\alias{aggregate_mzroll_nest}
\title{Aggregate mzroll lists}
\usage{
aggregate_mzroll_nest(mzroll_list_nest, samples_tbl)
}
\arguments{
\item{mzroll_list_nest}{a nested list of mzroll_lists produced from
\link{process_mzroll}.}

\item{samples_tbl}{Table of sample metadata}
}
\value{
a single mzroll_list
}
\description{
Combine mzroll lists into a single mzroll_list
}
