% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{augment_samples_with_samplesheet}
\alias{augment_samples_with_samplesheet}
\title{Augment Samples with Samplesheet}
\usage{
augment_samples_with_samplesheet(samples, sample_sheet_list)
}
\arguments{
\item{samples}{samples table generated within \link{process_mzroll}.}

\item{sample_sheet_list}{list of googlesheets information describing
experimental design - generated with \link{import_sample_sheet}.}
}
\value{
samples with added metadata from sample_sheet_list
Issue 297: This block was adjusted, and ultimately reverted
  Make sure there are only 1-1 substring matches between samples im
  dataset and sample sheet

e. g. if samples are named "C_200_B.mzML" and "C_2.mzML", "C_2"
  as an MS ID string will match to both samples.
To avoid this problem in this case, the MS ID string could be
  "C_2." or "C_2.mzML" instead of "C_2".
}
\description{
Augment Samples with Samplesheet
}
