% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{calculate_pathway_enrichment}
\alias{calculate_pathway_enrichment}
\title{Calculate Pathway Enrichments}
\usage{
calculate_pathway_enrichment(term_data, pathway_list, padj_cutoff = 0.2)
}
\arguments{
\item{term_data}{test statistics for a single term}

\item{pathway_list}{a named list formatted for fgsea with a name
corresponding to a pathway and entries matching the analytes included
in the set.}

\item{padj_cutoff}{minimum adjusted p-value to use for creating a the
pathway summary plot.}
}
\value{
a list containing a summary plot for the term and tibble
  containing data and plots for all comparisons performed.
}
\description{
Calculate Pathway Enrichments
}
