% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{calculate_pathway_enrichment_gsea}
\alias{calculate_pathway_enrichment_gsea}
\title{Calculate Pathway Enrichments using GSEA}
\usage{
calculate_pathway_enrichment_gsea(term_data, pathway_list, fdr_cutoff = 0.1)
}
\arguments{
\item{term_data}{test statistics for a single term}

\item{pathway_list}{a named list where names are pathways and entries
are groupIds belonging to each pathway.}

\item{fdr_cutoff}{minimum adjusted p-value to use for creating a the
pathway summary plot.}
}
\value{
a list containing a summary plot for the term and tibble
  containing data and plots for all comparisons performed.
}
\description{
Calculate Pathway Enrichments using GSEA
}
