% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mzroll_list.R
\name{check_infinite_values}
\alias{check_infinite_values}
\alias{check_peaks_infinite}
\alias{check_peaks_NA}
\title{Check groupIds for infinite or NA values}
\usage{
check_infinite_values(
  mzroll_list,
  infinite_type = "infinite",
  quant_var = "centered_log2_abundance",
  threshold = 0,
  filter_var = NULL,
  filter_ids = NULL,
  verbose = TRUE
)

check_peaks_infinite(mzroll_list, quant_var, threshold)

check_peaks_NA(mzroll_list, quant_var, threshold)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
\link{process_mzroll_multi}}

\item{infinite_type}{one of "infinite" or "NA" to indicate whether to check
for NA values or non-NA infinite (Inf, -Inf, and NaN) values}

\item{quant_var}{variable in measurements to use for abundance}

\item{threshold}{threshold above which groupIds will be returned. A value of
0 is recommended for \code{infinite_type = "infinite"} to return _any_
groupId containing any infinite value. A higher value or a value of 1 is
recommended for \code{infinite_type = "NA"} to return groupIds with a high
fraction or complete fraction of NA values.}

\item{filter_var}{column name on which to filter \code{filter_ids}; must be
a column name present in the \code{samples} or \code{features} tables}

\item{filter_ids}{sample or peak type ids on which to filter data from column
provided by \code{filter_var}}

\item{verbose}{when \code{TRUE} prints the compoundNames of groupIds that
exceed \code{threshold}}
}
\value{
a character vector of groupIds recommended for checking or filtering
}
\description{
Extract groupIds and optionally print compoundNames of peaks containing
non-finite values
}
