% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mzroll_list.R
\name{check_outliers}
\alias{check_outliers}
\title{Find outliers from numerical data columns}
\usage{
check_outliers(
  df,
  outlier_col_names = c("PC1", "PC2", "PC3"),
  outlier_sd = 3,
  sample_name_col = "sample_name"
)
}
\arguments{
\item{df}{a dataframe containing, at minimum, columns corresponding to
\code{outlier_col_names} and \code{sample_name_col}}

\item{outlier_col_names}{column name(s) as string or character vector on
which to calculate mean and standard deviation}

\item{outlier_sd}{standard deviation above or below which a sample will
be flagged as an outlier}

\item{sample_name_col}{column name as string of outlier sample names to return}
}
\value{
a character vector of \code{sample_name_col} values that were
identified as outliers
}
\description{
This function determines outliers and returns the sample names of outliers
based on values in \code{outlier_col_names} that are \code{outlier_sd}
(default = 3) standard deviations above or below the mean for all columns
provided. The default columns are "PC1", "PC2", and "PC3" from principal
component analysis.

Ported from \code{metstats} and used for \link[claman]{normalize_peaks_lm}
}
