% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_injections}
\alias{collapse_injections}
\title{Collapse Injections}
\usage{
collapse_injections(
  mzroll_list,
  grouping_vars,
  peak_quant_vars,
  collapse_fxn = "mean"
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
\link{process_mzroll_multi}}

\item{grouping_vars}{character vector of sample variables to use for
grouping}

\item{peak_quant_vars}{character vector of quantification variables to
average}

\item{collapse_fxn}{function name to use for collapse}
}
\value{
a \code{\link{process_mzroll}} collapsed across grouping_vars
}
\description{
Collapse Injections
}
\examples{
collapse_injections(
  nplug_mzroll_augmented,
  grouping_vars = "condition",
  peak_quant_vars = "log2_abundance"
)
}
