% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_metabolites}
\alias{collapse_metabolites}
\title{Collapse Metabolites}
\usage{
collapse_metabolites(
  mzroll_list,
  preserve_distinct_methods = FALSE,
  preserve_adducts = FALSE
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{preserve_distinct_methods}{if TRUE then collapse metabolites
for each method separately. If FALSE, collapse over methods.}

\item{preserve_adducts}{if TRUE then different ions of the same
metabolite will not be collapsed.}
}
\value{
an mzroll_list
}
\description{
Reduce multiple measurements of a metabolites within or across methods
  into a single consensus.
}
\examples{
collapse_metabolites(nplug_mzroll_augmented)

}
