% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_db_access}
\alias{configure_db_access}
\title{Configure Database Access}
\usage{
configure_db_access(standards_db = "standard")
}
\arguments{
\item{standards_db}{type of standards database to connect to:
\itemize{
  \item{standard: just standards}
  \item{extended: standards, external libraries and predicted spectra}
  }}
}
\value{
a list containing connections to the standards and systematic
  compounds database
}
\description{
Setup and test connection to Calico databases
}
