% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mzroll_list.R
\name{expand_peaks}
\alias{expand_peaks}
\title{Expand all samples and features}
\usage{
expand_peaks(
  mzroll_list,
  quant_var = "log2_abundance",
  log2_floor_value = NULL
)
}
\arguments{
\item{mzroll_list}{an \code{mzroll_list} tomic object}

\item{quant_var}{peak measurement variable on which to determine missing
values}

\item{log2_floor_value}{minimum value to set for missing peaks only. Will
also accept "min" or "halfmin" to impute NA values with groupId-specific
minimum or half-minimum values, respectively}
}
\value{
an mzroll_list object with groupId and sampleId values expanded
}
\description{
This function takes an mzroll_list (mzroll_tomic) object and expands
the features and samples in the measurements dataframe. All original peak
area values are retained. If \code{log2_floor_value} is not provided, all
missing values are enumerated with NA values. If a \code{log2_floor_value} is
provided as a numeric value, all NA values will be replaced with that value.
The \code{log2_floor_value} argument can also accept strings "min" or
"halfmin" which will replace NA values with groupId-specific log2 minimum or
half-minimum values, respectively.
}
\details{
This function differs from \code{claman::floor_peaks} as values below the
\code{log2_floor_value} are _not_ replaced with the floor values.
}
