% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mzroll_list.R
\name{extract_ids_from_metadata}
\alias{extract_ids_from_metadata}
\title{Extract groupIds or sampleIds per metadata}
\usage{
extract_ids_from_metadata(mzroll_list, filter_var, filter_ids)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
\link{process_mzroll_multi}}

\item{filter_var}{column name on which to filter \code{filter_ids}; must be
a column name present in the \code{samples} or \code{features} tables}

\item{filter_ids}{sample or peak type ids on which to filter data from column
provided by \code{filter_var}}
}
\value{
a list of 2 objects, where the first object is one of "sampleId" or
"groupId" and the second is a character vector of associated Ids
}
\description{
For a given filter_var and filter_ids, extract the associated groupIds and
sampleIds for easy filtering
}
