% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mzroll_list.R
\name{filter_samples_quo}
\alias{filter_samples_quo}
\title{Filter Calicomics Samples}
\usage{
filter_samples_quo(mzroll_list, filter_quosure)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{peakgroups: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{peaks: one row per peak (samples x peakgroups)}
  }}

\item{filter_quosure}{quosure to use to filter samples}
}
\value{
an \code{mzroll_list} with samples (and corresponding peaks) removed
}
\description{
Remove samples based on sample table variables
}
\examples{
mzroll_list <- readRDS(authutils::get_clamr_assets("X0083-mzroll-list.Rds"))
filter_quosure <- rlang::quo(sex == "Male")
filter_samples_quo(mzroll_list, filter_quosure)
}
