% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mzroll_list.R
\name{floor_peaks}
\alias{floor_peaks}
\title{Floor Peaks}
\usage{
floor_peaks(mzroll_list, log2_floor_value = 12, floor_var = "log2_abundance")
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
\link{process_mzroll_multi}}

\item{log2_floor_value}{minimum value to set for low abundance or
missing peaks}

\item{floor_var}{measurement variable to floor to \code{log2_floor_value}.}
}
\value{
\code{\link{process_mzroll}}
}
\description{
Set a minimum peak abundance of floor_value for low abundance and
  undetected peaks.
}
\examples{
floored_peaks <- floor_peaks(nplug_mzroll_augmented, 12)
}
