% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{hclust_order}
\alias{hclust_order}
\title{Hierarchical clustering order}
\usage{
hclust_order(df, feature_pk, sample_pk, measurement_var, cluster_dim)
}
\arguments{
\item{df}{data.frame to cluster}

\item{feature_pk}{variable uniquely defining a row}

\item{sample_pk}{variable uniquely definining a sample}

\item{measurement_var}{measurement variables}

\item{cluster_dim}{rows, columns, or both}
}
\value{
a list containing a hierarchically clustered set of rows
  and/or columns
}
\description{
Format and hierarchically cluster a data.frame. If hclust could not
  normally be produced (usually because no samples are in common for a
  feature) pad the matrix with zeros and still calculate the distance
}
