% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{import_sample_sheet}
\alias{import_sample_sheet}
\title{Import Sample List}
\usage{
import_sample_sheet(pattern)
}
\description{
Find, read and process a sample meta sheet from googlesheets
}
\examples{
\dontrun{
import_sample_sheet(pattern = "X0083")
}

}
