% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mzroll_list.R
\name{impute_missing_peaks}
\alias{impute_missing_peaks}
\title{Impute missing peaks with provided feature-level imputation values}
\usage{
impute_missing_peaks(
  mzroll_list,
  lod_values,
  quant_var = "log2_abundance",
  imputation_sd = 0.15
)
}
\arguments{
\item{mzroll_list}{data in triple omic structure}

\item{lod_values}{a tibble that maps groupId to log2 feature-level imputation values}

\item{quant_var}{column to use for peak values}

\item{imputation_sd}{standard deviation of Gaussian distribution to use for missing peak imputation}
}
\value{
triple omic data with imputed missing peaks
}
\description{
Impute missing peaks with provided feature-level imputation values
}
\examples{
library(dplyr)

lod_values <- nplug_mzroll_augmented[["measurements"]] \%>\%
  dplyr::select(groupId, log2_abundance) \%>\%
  dplyr::distinct(groupId, .keep_all = TRUE)

mzroll_list <- nplug_mzroll_augmented
mzroll_list$measurements <- mzroll_list$measurements \%>\% filter(groupId != 2 & sampleId != 1)

mzroll_list_imputed <- impute_missing_peaks(mzroll_list, lod_values, "log2_abundance", 0.15)

}
