% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_has_label}
\alias{is_has_label}
\title{Util - Check if Peak Group matches label filters}
\usage{
is_has_label(peakgroup_labels, labels_to_keep = "*", labels_to_exclude = "")
}
\arguments{
\item{peakgroup_labels}{String encoding of labels.
Each character in the string corresponds to a different label.}

\item{labels_to_keep}{Retain the peakgroup if any 
character in the string matches to any character in this string.
\code{default = "*"} (retain all labels)}

\item{labels_to_exclude}{Discard the peakgroup if any 
character in the string matches to any character in this string. 
If a label is both flagged for inclusion and exclusion, it is excluded.
\code{default = ""} (do not exclude any labels)}
}
\value{
TRUE or FALSE, if the peakgroup label matches the criteria.
}
\description{
Filter peak groups based on matching labels.
  Labels are encoded as string, but each character corresponds to a different label.
  This filter operates as an "or" filter on both the set of labels to keep or exclude.
  To specify a peakgroup that has no labels, use a single white space in the string (\code{' '}).
}
