% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lipid_components}
\alias{lipid_components}
\title{Generate relevant lipid components from compound name}
\usage{
lipid_components(compound_name)
}
\arguments{
\item{compound_name}{vector of compound names}
}
\value{
compound_name vector expanded to large table, with structural
  components described as their own columns.
}
\description{
Produce data frame with lipid components from name
}
