% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mzroll_list.R
\name{median_polish_predict_dilutions}
\alias{median_polish_predict_dilutions}
\title{Predict Dilutions from Median Polish Scaling Factor}
\usage{
median_polish_predict_dilutions(
  mzroll_list,
  scaling_factor = "median_polish_scaling_factor",
  norm_scale_varname = "median_polish_predicted_dilutions",
  group_vars = NULL
)
}
\arguments{
\item{mzroll_list}{data in triple omic structure}

\item{scaling_factor}{scaling factor, defaults to `median_polish_scaling_factor`
and must be a `samples` column in `mzroll_list`}

\item{norm_scale_varname}{variable in samples to add for dilution predictions}

\item{group_var}{optional grouping variable on which to calculate maximum dilution,
must be a `samples` column in `mzroll_list`}
}
\value{
a \code{mzroll_list} with \code{norm_scale_varname} variable added to 
samples
}
\description{
Using `median_polish_scaling_factor` output from `normalize_peaks_median_polish`,
predict sample-wise dilutions
}
\details{
This function performs an inverse-log transformation on the scaling factors,
and scales the samples based on the observed maximum scaling factor. Assumes that
median polish is perform on log2 transformed data
}
