% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{merge_samples_tbl}
\alias{merge_samples_tbl}
\title{Merge Samples Table}
\usage{
merge_samples_tbl(mzroll_list, samples_tbl, id_strings, exact = FALSE)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
\link{process_mzroll_multi}}

\item{samples_tbl}{Table of sample metadata}

\item{id_strings}{one or more variables which will be used to match
sample names.}

\item{exact}{if true, an exact match between mzroll names and id_strings
will be found; if false, then substring matches will be used.}
}
\description{
Merge a table of sample metadata with an existing mzroll_list
}
\examples{
mzroll_list <- process_mzroll(nplug_mzroll())
merge_samples_tbl(mzroll_list, nplug_samples, "sample_name", exact = TRUE)
}
