% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mzroll_table}
\alias{mzroll_table}
\title{Flatten mzroll list to single table}
\usage{
mzroll_table(mzroll_list)
}
\arguments{
\item{mzroll_list}{output of process_mzrollDB functions
\itemize{
  \item{peakgroups: one row per unique analyte (defined by a unique
    groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{peaks: one row per peak (samples x peakgroups)}
  }}
}
\value{
mzroll_list (process_mzrollDB functions output) reformatted as a
  single table
}
\description{
Flatten mzroll list to single table using groupId and sampleId
}
