% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nplug_compounds}
\alias{nplug_compounds}
\title{NPLUG samples}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 108 rows and 2 columns.
}
\usage{
nplug_compounds
}
\description{
A table of meta-data for all NPLUG features with variables:
\describe{
  \item{compoundName}{compound name corresponding to the compoundName
    variable in the nplug_mzroll's features table.}
  \item{pathway}{A (rough) categorization of compounds into metabolic
    pathways.}
    }
}
\seealso{
Other nplug: 
\code{\link{nplug_mzroll_augmented}},
\code{\link{nplug_mzroll}()},
\code{\link{nplug_samples}}
}
\concept{nplug}
\keyword{datasets}
