% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nplug_mzroll}
\alias{nplug_mzroll}
\title{Growth-Limiting Metabolites}
\usage{
nplug_mzroll()
}
\value{
path to the NPLUG .mzrollDB dataset
}
\description{
The NPLUG experiment has a full factorial design exploring relationship
  between nutrient limitation and growth rate in yeast.
}
\details{
The experiment contains 25 primary experimental conditions:
\itemize{
  \item{5 limiting nutrients: nitrogen, phosphorous, leucine, uracil and
    glucose (i.e., NPLUG!)}
  \item{5 dilution rates (growth rate / ln(2)): from 0.05-0.3 1/hrs} 
}

Metabolomics of 
  limited yeast cultures growing at different rates.
}
\seealso{
Other nplug: 
\code{\link{nplug_compounds}},
\code{\link{nplug_mzroll_augmented}},
\code{\link{nplug_mzroll_normalized}},
\code{\link{nplug_samples}}
}
\concept{nplug}
