% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nplug_mzroll_augmented}
\alias{nplug_mzroll_augmented}
\title{NPLUG MzRoll Augmented}
\format{
An object of class \code{triple_omic} (inherits from \code{tomic}, \code{mzroll}) of length 4.
}
\usage{
nplug_mzroll_augmented
}
\description{
\link{nplug_mzroll} formatted with \link{process_mzroll} with sample
  (\link{nplug_samples}) and compound (\link{nplug_samples}) merged.
}
\seealso{
Other nplug: 
\code{\link{nplug_compounds}},
\code{\link{nplug_mzroll}()},
\code{\link{nplug_mzroll_normalized}},
\code{\link{nplug_samples}}
}
\concept{nplug}
\keyword{datasets}
