% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nplug_mzroll_normalized}
\alias{nplug_mzroll_normalized}
\title{NPLUG MzRoll Augmented}
\format{
An object of class \code{triple_omic} (inherits from \code{tomic}, \code{mzroll}) of length 4.
}
\usage{
nplug_mzroll_normalized
}
\description{
\link{nplug_mzroll_augmented} with injections collapsed using 
  \link{collapse_injections}, followed by reference-sample normalization
  using \link{normalize_peaks}. Finally, reference samples and samples
  extracting using the pellet method were removed and sample names
  were cleaned up. These steps are described in the NPLUG vignette.
}
\seealso{
Other nplug: 
\code{\link{nplug_compounds}},
\code{\link{nplug_mzroll_augmented}},
\code{\link{nplug_mzroll}()},
\code{\link{nplug_samples}}
}
\concept{nplug}
\keyword{datasets}
