% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nplug_samples}
\alias{nplug_samples}
\title{NPLUG samples}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 136 rows and 9 columns.
}
\usage{
nplug_samples
}
\description{
A table of meta-data for all NPLUG samples with variables:
\describe{
  \item{sample_name}{unique descriptor of each sample}
  \item{month}{month of sample generation}
  \item{replicate}{culture replicate}
  \item{DR}{dilution rate of culture - this is proportional to cells'
  growth rate}
  \item{limitation}{nutrient limiting growth
    \itemize{
      \item{GLU - carbon}
      \item{LEU - leucine (in a Leu4 auxotroph)}
      \item{NH4 - nitrogen}
      \item{PO4 - phosphorous}
      \item{URA - uracil (in a Ura2 auxotroph)}
    }
  }
  \item{exp_ref}{experimental or reference condition}
  \item{extraction}{filter- or pellet-based extraction}
  \item{condition}{Integer value for each unique condition. Here, that is
    a unique values of (month, limitation, DR, and extraction) for
    experimental samples, and unique values of (month and extraction) for
    reference samples.}
  \item{reference}{Condition number that this sample should be compared to.}
  }
}
\seealso{
Other nplug: 
\code{\link{nplug_compounds}},
\code{\link{nplug_mzroll_augmented}},
\code{\link{nplug_mzroll_normalized}},
\code{\link{nplug_mzroll}()}
}
\concept{nplug}
\keyword{datasets}
