% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Plot Barplots}
\usage{
plot_barplot(mzroll_list, groupIds)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{groupIds}{groupIds for compounds to plot}
}
\description{
Plot Barplots
}
