% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Omics Heatmap}
\usage{
plot_heatmap(
  mzroll_list,
  feature.var = "groupId",
  sample.var = "name",
  value.var = "centered_log2_abundance",
  cluster_dim = "both",
  change_threshold = 3,
  plot_type = "plotly"
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{feature.var}{variable from "peakgroups" to use as a unique feature
label.}

\item{sample.var}{variable from "samples" to use as a unique sample label.}

\item{value.var}{which variable in "peaks" to use for quantification.}

\item{cluster_dim}{dimensions to cluster (row, column, or both)}

\item{change_threshold}{values with a more extreme absolute change will be
thresholded to this value.}

\item{plot_type}{plotly (for interactivity) or grob (for a static ggplot)}
}
\description{
Generate a basic heatmap of all of your metabolomic and lipidomic data
}
\examples{
library(dplyr)
mzroll_list <- authutils::get_clamr_assets("X0083-M001A.mzrollDB") \%>\%
  process_mzroll(
    standard_databases = NULL,
    method_tag = "M001A",
    only_identified = TRUE
  ) \%>\%
  floor_peaks(12)

plot_heatmap(
  mzroll_list,
  sample.var = "sampleId",
  feature.var = "peak_label",
  change_threshold = 5,
  cluster_dim = "rows",
  plot_type = "grob"
)
\dontrun{
plot_heatmap(
  mzroll_list,
  sample.var = "sampleId",
  feature.var = "peak_label",
  change_threshold = 5,
  cluster_dim = "rows"
)
}

}
