% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{plot_pvalues}
\alias{plot_pvalues}
\title{P-value Histogram}
\usage{
plot_pvalues(regression_significance)
}
\arguments{
\item{regression_significance}{returned by \code{\link{diffex_mzroll}};
a tibble of tests performed.}
}
\description{
P-value Histogram
}
\examples{
regression_significance <- diffex_mzroll(
  nplug_mzroll_normalized,
  "normalized_log2_abundance",
  "limitation + limitation:DR + 0"
)

plot_pvalues(regression_significance)
}
