% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(regression_significance, max_p_trans = 10, FDR_cutoff = 0.1)
}
\arguments{
\item{regression_significance}{output of diffex_mzroll}

\item{max_p_trans}{maximum value of -log10 pvalues to plot}

\item{FDR_cutoff}{FDR cutoff to label for significance}
}
\value{
a grob
}
\description{
Volcano plot
}
\examples{
regression_significance <- diffex_mzroll(
  nplug_mzroll_normalized,
  "normalized_log2_abundance",
  "limitation + limitation:DR + 0"
  )

plot_volcano(regression_significance, 10, 0.1)

}
