% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{process_metabolon}
\alias{process_metabolon}
\title{Process Metabolon}
\usage{
process_metabolon(
  peak_areas_path,
  chem_anno_path,
  metadata_path,
  peak_floor_intensity = 4096
)
}
\arguments{
\item{peak_areas_path}{path to peak areas CSV file from metabolon results}

\item{chem_anno_path}{path to chemical annotations CSV file from metabolon results}

\item{metadata_path}{path to metadata CSV file from metabolon results}

\item{peak_floor_intensity}{value to use to replace NA peak measurements (missing values)}
}
\description{
Import results generated by metabolon into romic structure, using series of CSV
output files delivered by Metabolon as inputs.
}
