% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{process_mzroll_identify_peakgroups}
\alias{process_mzroll_identify_peakgroups}
\title{Process MzRoll - Identify Peakgroups}
\usage{
process_mzroll_identify_peakgroups(peakgroups, only_identified)
}
\arguments{
\item{peakgroups}{a table of distinct ions with characteristic m/z and rt}

\item{only_identified}{TRUE/FALSE, filter to only features which were
identified.}
}
\value{
a table of peakgroups
}
\description{
Either remove unidentified peakgroups or name unknowns using m/z and rt.
}
