% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{process_mzroll_validate_peakgroups}
\alias{process_mzroll_validate_peakgroups}
\title{Process MzRoll - Validate Peakgroups}
\usage{
process_mzroll_validate_peakgroups(peakgroups, validate)
}
\arguments{
\item{peakgroups}{a table of distinct ions with characteristic m/z and rt}

\item{validate}{TRUE/FALSE, use meta-data to only name the subset of
features which were manually validated.}
}
\value{
a table of peakgroups
}
\description{
If validate is True then remove unvalidated compoundNames (based on
  peakgroup labels) and identify all unvalidated compounds with an
  "is_unknown" variable.
}
