% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{pvalue_histograms}
\alias{pvalue_histograms}
\title{P-value Histogram}
\usage{
pvalue_histograms(regression_significance)
}
\arguments{
\item{regression_significance}{output of diffex_mzroll}
}
\description{
P-value Histogram
}
