% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_db_query.R
\name{query_systematic_compounds}
\alias{query_systematic_compounds}
\title{Query Systematic Compounds}
\usage{
query_systematic_compounds(
  systematicCompoundIds = NULL,
  systematic_compounds_con
)
}
\arguments{
\item{systematicCompoundIds}{NULL to query all compounds or a numeric
vector of systematic compound IDs}

\item{systematic_compounds_con}{A connection to an SQL database which
stores the structure and systematic IDs of well-studied compounds.}
}
\value{
a list containing:
\itemize{
  \item{distinct_compounds: unique database entries}
  \item{database_ids: systematic IDs of compounds in KEGG, HMDB, ...}
  \item{compound_aliases: alternative names for a compound}
  \item{metabolic_pathways: KEGG pathways associated with a metabolite}
  }
}
\description{
Query Systematic Compounds
}
\examples{
\dontrun{
systematic_compounds_con <- configure_db_access()$systematic_compounds_con
systematic_compounds <- query_systematic_compounds(
  systematic_compounds_con = systematic_compounds_con
)
}
}
