% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{read_sample_list}
\alias{read_sample_list}
\title{Read Sample List}
\usage{
read_sample_list(tracking_sheet_id)
}
\arguments{
\item{tracking_sheet_id}{output of \link{find_tracking_sheet}}
}
\description{
Read Sample List
}
