% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{read_samples_tbl}
\alias{read_samples_tbl}
\title{Read Samples Table}
\usage{
read_samples_tbl(samples_tbl)
}
\arguments{
\item{samples_tbl}{Table of sample metadata}
}
\description{
Read Samples Table
}
