% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mzroll_list.R
\name{reconcile_mzroll_list}
\alias{reconcile_mzroll_list}
\title{Reconcile Mzroll List}
\usage{
reconcile_mzroll_list(mzroll_list)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{peakgroups: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{peaks: one row per peak (samples x peakgroups)}
  }}
}
\description{
If some samples, peakgroups or peaks have been dropped; peaks and peakgroups
  which may now be irrelevant
}
