% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{remove_constant_name}
\alias{remove_constant_name}
\title{Remove Constant Name}
\usage{
remove_constant_name(name_set)
}
\arguments{
\item{name_set}{a character vector of names with some common leading or
lagging substrings.}
}
\value{
name_set with constant substring removed.
}
\description{
Taking a set of filenames remove the constant leading &/or lagging portion
  of the names.
}
\examples{
remove_constant_name(c("xxxxsample1yyyy", "xxxxcontrol1yyyy"))
}
