% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_db_query.R
\name{summarize_compound_pathways}
\alias{summarize_compound_pathways}
\title{Summarize compound pathways}
\usage{
summarize_compound_pathways(
  systematic_compounds,
  min_pw_size = 5L,
  focus_pathways = c("Glycolysis / Gluconeogenesis", "Citrate cycle (TCA cycle)",
    "Pentose phosphate pathway", "Biosynthesis of amino acids", "Purine metabolism",
    "Pyrimidine metabolism")
)
}
\arguments{
\item{systematic_compounds}{The output of
\code{\link{query_systematic_compounds}}.}

\item{min_pw_size}{Minimum number of examples of metabolites from a pathway
for the pathway not to be filtered.}

\item{focus_pathways}{A character vector of names of KEGG pathways to be
included as \code{focus_pathway} annotations.}
}
\value{
a tibble of systematicCompoundIds and pathway annotations
}
\description{
Reduce compound pathway annotations to a set of core "focus" pathways and
  to well-represented (>= \code{min_pw_size} examples) general pathways.
}
