% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_mzroll_list}
\alias{test_mzroll_list}
\title{Test Mzroll List}
\usage{
test_mzroll_list(mzroll_list)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{peakgroups: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{peaks: one row per peak (samples x peakgroups)}
  }}
}
\description{
Test Mzroll List
}
