% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_pretty_khead}
\alias{util_pretty_khead}
\title{Util - Pretty Knitr Head}
\usage{
util_pretty_khead(tbl, nrows = 10, caption = NULL)
}
\arguments{
\item{tbl}{a data.frame or tibble}

\item{nrows}{the max number of rows to show}

\item{caption}{The table caption. By default, it is retrieved from the chunk
option \code{tab.cap}.}
}
\value{
an html knitr table
}
\description{
Util - Pretty Knitr Head
}
\examples{
util_pretty_khead(mtcars, nrows = 5, caption = "cars!")
}
