% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(regression_significance, max_p_trans = 10, FDR_cutoff = 0.1)
}
\arguments{
\item{regression_significance}{output of diffex_mzroll}

\item{max_p_trans}{maximum value of -log10 pvalues to plot}

\item{FDR_cutoff}{FDR cutoff to label for significance}
}
\description{
Volcano plot
}
