% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{write_tidy_list_augmented}
\alias{write_tidy_list_augmented}
\title{Write flat tidy list}
\usage{
write_tidy_list_augmented(mzroll_list, dir_path, name_preamble)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{dir_path}{path to save outputs}

\item{name_preamble}{start of output file name}
}
\value{
Export one table which is one row per peak, which includes all feature and
  sample attributes.
}
\description{
Write flat tidy list
}
