% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{write_tidy_list_wide}
\alias{write_tidy_list_wide}
\title{Write wide output}
\usage{
write_tidy_list_wide(
  mzroll_list,
  dir_path,
  name_preamble,
  value.var = "log2_abundance",
  transpose = FALSE
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{dir_path}{path to save outputs}

\item{name_preamble}{start of output file name}

\item{value.var}{measurement variable}

\item{transpose}{if TRUE then samples will be stored as rows}
}
\value{
Export one table which contains metabolites as rows and samples as columns.
}
\description{
abundances form a matrix with metabolites as rows and samples as columns.
  Use transpose to treat samples as rows
}
