% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_metabolites}
\alias{collapse_metabolites}
\title{Collapse Metabolites}
\usage{
collapse_metabolites(
  mzroll_list,
  preserve_distinct_methods = FALSE,
  preserve_adducts = FALSE
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{preserve_distinct_methods}{if TRUE then collapse metabolites
for each method separately. If FALSE, collapse over methods.}

\item{preserve_adducts}{if TRUE then different ions of the same
metabolite will not be collapsed.}
}
\value{
an mzroll_list
}
\description{
Combine multiple measurements of the same metabolite into a consensus
  to simplify result's presentation and pathway analysis.
}
\details{
Analytes are first aggregated by retaining the maximum
  intensity peak on a sample-by-sample basis over peakgroups of the
  same ion (i.e., same compoundName and adductName). This is meant to
  deal with peakgroup splitting. Once measurements are reduced to
  unique ions, ions can be further aggregated to metabolites by
  taking the median quant value on a sample-by-sample basis while
  preserving either adducts or methods.
}
\examples{
collapse_metabolites(nplug_mzroll_augmented)

}
