% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{diffex_mzroll}
\alias{diffex_mzroll}
\title{Differential Expression Analysis on Mzroll List}
\usage{
diffex_mzroll(mzroll_list, value_var, test_model, null_model = NULL)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{value_var}{measurement variable}

\item{test_model}{a RHS formula for regression}

\item{null_model}{if provided a null RHS formula to compare to
\code{test_model} using the likelihood-ratio test.}
}
\value{
a tibble of significance tests for each feature
}
\description{
Differential Expression Analysis on Mzroll List
}
\examples{
diffex_mzroll(
  nplug_mzroll_normalized,
  "normalized_log2_abundance",
  "limitation + limitation:DR + 0"
  )
  
diffex_mzroll(
  nplug_mzroll_normalized,
  "normalized_log2_abundance",
  "limitation + 0",
  "0"
  )
}
