% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{diffex_mzroll}
\alias{diffex_mzroll}
\title{Differential Expression Analysis on Mzroll List}
\usage{
diffex_mzroll(
  mzroll_list,
  value.var = "centered_log2_abundance",
  test_model,
  null_model = NULL
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{value.var}{measurement variable}

\item{test_model}{a RHS formula for regression}

\item{null_model}{if provided a null RHS formula to compare to
\code{test_model} using the likelihood-ratio test.}
}
\value{
a tibble of significance tests for each feature
}
\description{
Differential Expression Analysis on Mzroll List
}
\examples{
library(dplyr)

mzroll_list <- readRDS(
  authutils::get_clamr_assets("X0083-mzroll-list.Rds")
) \%>\%
  floor_peaks(12)

diffex_mzroll(mzroll_list, test_model = "mutation + sex")
}
