% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mzroll_list.R
\name{fill_in_missing_peaks}
\alias{fill_in_missing_peaks}
\title{Fill in missing peaks}
\usage{
fill_in_missing_peaks(
  mzroll_list,
  fill_values,
  quant_var = "log2_abundance",
  imputation_sd = 0.15
)
}
\arguments{
\item{mzroll_list}{data in triple omic structure}

\item{fill_values}{either a numeric constant or a tibble that maps groupId to log2 feature-level imputation values}

\item{quant_var}{column to use for peak values}

\item{imputation_sd}{standard deviation of Gaussian distribution to use for missing peak imputation}
}
\value{
triple omic data with imputed missing peaks
}
\description{
If \code{fill_values} is a data frame, this function calls \code{impute_missing_peaks()}.
If it is a numeric vector, 'this function calls \code{floor_peaks()}. Other types are not currently supported
}
\examples{
library(dplyr)

lod_values <- nplug_mzroll_augmented[["measurements"]] \%>\%
dplyr::select(groupId, log2_abundance) \%>\%
dplyr::distinct(groupId, .keep_all = TRUE)

mzroll_list <- nplug_mzroll_augmented
mzroll_list$measurements <- mzroll_list$measurements \%>\% filter(groupId != 2 & sampleId != 1)

mzroll_list_imputed <- fill_in_missing_peaks(mzroll_list, lod_values, "log2_abundance", 0.15)
mzroll_list_imputed <- fill_in_missing_peaks(mzroll_list, 12, "log2_abundance")

}
