% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mzroll_list.R
\name{filter_groupIds}
\alias{filter_groupIds}
\title{Filter GroupIds}
\usage{
filter_groupIds(mzroll_list, groupIds, invert = FALSE)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{groupIds}{groupIds to retain}

\item{invert}{if TRUE then remove provided groupIds; if FALSE then retain
provided groupIds}
}
\description{
Remove all data besides those from provided groupIds
}
\examples{
filter_groupIds(nplug_mzroll_normalized, factor(1:10))
}
