% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mzroll_list.R
\name{floor_peaks}
\alias{floor_peaks}
\title{Floor Peaks}
\usage{
floor_peaks(mzroll_list, log2_floor_value = 12)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{log2_floor_value}{minimum value to set for low abundance or
missing peaks}
}
\value{
\code{\link{process_mzroll}}
}
\description{
Set a minimum peak abundance of floor_value for low abundance and
  undetected peaks.
}
