% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds_metadata.R
\name{merge_compounds_tbl}
\alias{merge_compounds_tbl}
\title{Merge Compounds Table}
\usage{
merge_compounds_tbl(mzroll_list, compounds_tbl, by = "compoundName")
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
\link{process_mzroll_multi}}

\item{compounds_tbl}{Table of compound metadata}

\item{by}{variable shared by mzroll peakgroups and compounds_tbl
to merge results by.}
}
\description{
Merge a table of compound information with an existing mzroll_list
}
\examples{
mzroll_list <- process_mzroll(nplug_mzroll())
compounds_tbl <- nplug_compounds
merge_compounds_tbl(mzroll_list, compounds_tbl)
}
