% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mzroll_list.R
\name{normalize_peaks}
\alias{normalize_peaks}
\alias{normalize_peaks_median_polish}
\alias{normalize_peaks_loading_value}
\alias{normalize_peaks_batch_center}
\alias{normalize_peaks_reference_sample}
\alias{normalize_peaks_reference_condition}
\alias{normalize_peaks_loess}
\alias{normalize_peaks_lm}
\alias{normalize_peaks_center}
\title{Normalize Peaks}
\usage{
normalize_peaks(
  mzroll_list,
  normalization_method,
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_IC",
  ...
)

normalize_peaks_median_polish(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  log2_floor_value = NA
)

normalize_peaks_loading_value(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  loading_varname,
  log2_floor_value = NA
)

normalize_peaks_batch_center(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  batch_varnames,
  log2_floor_value = NA,
  centering_fxn = mean
)

normalize_peaks_reference_sample(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  batch_varnames,
  reference_varname = "sample",
  reference_values = "posctl",
  log2_floor_value = NA
)

normalize_peaks_reference_condition(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  condition_varname = "condition #",
  reference_varname = "reference condition #",
  log2_floor_value = NA
)

normalize_peaks_loess(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  weights_tribble = NULL,
  log2_floor_value = 12
)

normalize_peaks_lm(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  time_col_varname
)

normalize_peaks_center(mzroll_list, quant_peak_varname, norm_peak_varname)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{normalization_method}{Normalization method to apply
\itemize{
  \item{\code{median polish}: column normalization based on average signal}
  \item{\code{loading value}: column normalization using a sample-level
    value
    }
  \item{\code{center batches}: batch centering}
  \item{\code{center}: zero-center each compound}
  \item{\code{reference sample}: compare to reference sample}
  \item{\code{reference condition}: compare each sample to its specified
    reference sample}
  \item{\code{loess}: weighted smoothing of IC over time (adds
    \code{.loess_fit} as a peaks variable in addition to
    \code{norm_peak_varname})}
  \item{\code{lm}: linear regression smoothing of IC over time (adds
    \code{lm_estimate} as feature variable in addition to
    \code{norm_peak_varname})}
  }}

\item{quant_peak_varname}{variable in measurements to use for abundance}

\item{norm_peak_varname}{variable in measurements to add for normalized
abundance}

\item{...}{additional arguments to pass to normalization methods}

\item{log2_floor_value}{minimum value to set for low abundance or
missing peaks}

\item{loading_varname}{sample variable used for adjustment}

\item{batch_varnames}{variables to use for grouping when removing batch
effects}

\item{centering_fxn}{function to use when centering (mean, median, ...)}

\item{reference_varname}{variable specifying which samples are references}

\item{reference_values}{values of \code{reference_varname} indicating
reference samples}

\item{condition_varname}{variable specifying each sample's condition}

\item{weights_tribble}{a table containing weights and sample variables to
match them to.}

\item{time_col_varname}{variable in samples table to use for linear 
correction}
}
\value{
a \code{mzroll_list} as generated by \code{\link{process_mzroll}}
  a \code{norm_peak_varname} variable added to measuremnets
}
\description{
Using a robust metabolomics normalization rescale ion counts based on the
  consensus signal of each sample.

Using a sample-level summary, such as number of cells, adjust all values.

Compare each sample to the reference samples within the same batch

Compare each sample to the reference samples within the same batch
}
\details{
The robust median polish was reported in Kamphorst et al. 2015, see
  \url{https://github.com/shackett/Pancreatic_tumor_metabolomics}.

Note, log2 intensities will be shifted down by this value, so
  make sure the provided values are appropriately transformed.

reference_varname specifies which condition to contrast a given sample to
  (using the condition values from condition_varname)
}
\examples{

normalize_peaks(
  nplug_mzroll_augmented,
  "median polish",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance"
)

# this examples doesn't make biological sense but its syntactically correct
normalize_peaks(
  nplug_mzroll_augmented,
  "loading value",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance",
  loading_varname = "DR"
)

normalize_peaks(
  nplug_mzroll_augmented,
  "center batches",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance",
  batch_varnames = "month",
  centering_fxn = median
)

normalize_peaks(
  nplug_mzroll_augmented,
  normalization_method = "reference condition",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance",
  condition_varname = "condition",
  reference_varname = "reference"
)

normalize_peaks(
  nplug_mzroll_augmented,
  normalization_method = "reference sample",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance",
  batch_varnames = c("month", "extraction"),
  reference_varname = "exp_ref",
  reference_values = "ref"
)
}
