% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mzroll_list.R
\name{normalize_peaks}
\alias{normalize_peaks}
\alias{normalize_peaks.median_polish}
\alias{normalize_peaks.batch_center}
\alias{normalize_peaks.reference_sample}
\alias{normalize_peaks.reference_condition}
\alias{normalize_peaks.loess}
\title{Normalize Peaks}
\usage{
normalize_peaks(
  mzroll_list,
  normalization_method,
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_IC",
  ...
)

normalize_peaks.median_polish(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  log2_floor_value = NA
)

normalize_peaks.batch_center(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  batch_varnames,
  log2_floor_value = NA,
  centering_fxn = mean
)

normalize_peaks.reference_sample(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  batch_varnames,
  reference_varname = "sample",
  reference_values = "posctl",
  log2_floor_value = NA
)

normalize_peaks.reference_condition(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  condition_varname = "condition #",
  reference_varname = "reference condition #",
  log2_floor_value = NA
)

normalize_peaks.loess(
  mzroll_list,
  quant_peak_varname,
  norm_peak_varname,
  weights_tribble = NULL,
  log2_floor_value = 12
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{normalization_method}{Normalization method to apply
\itemize{
  \item{\code{median polish}: column normalization}
  \item{\code{center batches}: batch centering}
  \item{\code{reference sample}: compare to reference sample}
  \item{\code{reference condition}: compare each sample to its specified
    reference sample}
  \item{\code{loess}: weighted smoothing of IC over time (adds
    \code{.loess_fit} as a peaks variable in addition to
    \code{norm_peak_varname})}
  }}

\item{quant_peak_varname}{variable in measurements to use for abundance}

\item{norm_peak_varname}{variable in measurements to add for normalized
abundance}

\item{...}{additional arguments to pass to normalization methods}

\item{log2_floor_value}{minimum value to set for low abundance or
missing peaks}

\item{batch_varnames}{variables to use for grouping when removing batch
effects}

\item{centering_fxn}{function to use when centering (mean, median, ...)}

\item{reference_varname}{variable specifying which samples are references}

\item{reference_values}{values of \code{reference_varname} indicating
reference samples}

\item{condition_varname}{variable specifying each sample's condition}

\item{weights_tribble}{a table containing weights and sample variables to
match them to.}
}
\value{
a \code{mzroll_list} as generated by \code{\link{process_mzroll}}
  a \code{norm_peak_varname} variable added to measuremnets
}
\description{
Using a robust metabolomics normalization rescale ion counts based on the
  consensus signal of each sample.

Compare each sample to the reference samples within the same batch

Compare each sample to the reference samples within the same batch
}
\details{
The robust median polish was reported in Kamphorst et al. 2015, see
  \url{https://github.com/shackett/Pancreatic_tumor_metabolomics}.

reference_varname specifies which condition to contrast a given sample to
  (using the condition values from condition_varname)
}
\examples{

normalize_peaks(
  nplug_mzroll_augmented,
  "median polish",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance"
)

normalize_peaks(
  nplug_mzroll_augmented,
  "center batches",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance",
  batch_varnames = "month",
  centering_fxn = median
)

normalize_peaks(
  nplug_mzroll_augmented,
  normalization_method = "reference condition",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance",
  condition_varname = "condition",
  reference_varname = "reference"
)

normalize_peaks(
  nplug_mzroll_augmented,
  normalization_method = "reference sample",
  quant_peak_varname = "log2_abundance",
  norm_peak_varname = "normalized_log2_abundance",
  batch_varnames = c("month", "extraction"),
  reference_varname = "exp_ref",
  reference_values = "ref"
)

}
