% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{pathway_enrichments}
\alias{pathway_enrichments}
\title{Pathway Enrichments}
\usage{
pathway_enrichments(
  mzroll_list,
  significance,
  standard_databases,
  test_absolute_effects = TRUE
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{peakgroups: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{peaks: one row per peak (samples x peakgroups)}
  }}

\item{significance}{returned by \code{\link{diffex_mzroll}}; a tibble of
tests performed.}

\item{test_absolute_effects}{If TRUE then only consider the magnitude of
test-statistics when calculating rankings for enrichment; if FALSE then
consider sign as well.}
}
\value{
a list containing:
\itemize{
  \item{enrichment_table: a tibble containing each term's enrichment for
    every pathway and an enrichment_plot}
  \item{enrichment_plots: pre-generated plots containing the most
    significant enrichments for each term}
  }
}
\description{
Pathway Enrichments
}
