% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Plot Barplots}
\usage{
plot_barplot(mzroll_list, groupIds, grouping_vars, value_var, fill_var = NULL)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
  \link{process_mzroll_multi}

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}

\item{groupIds}{groupIds for compounds to plot}

\item{grouping_vars}{variables to use for ordering samples and calculating
standard errors}

\item{value_var}{measurement variable}

\item{fill_var}{variable to use for filling bars}
}
\description{
Plot Barplots
}
\examples{

plot_barplot(
  mzroll_list = nplug_mzroll_normalized,
  groupIds = 1:4,
  grouping_vars = c("limitation", "DR"),
  value_var = "normalized_log2_abundance",
  fill_var = "limitation"
)
}
