% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{plot_compare_injection}
\alias{plot_compare_injection}
\title{Compare Injections}
\usage{
plot_compare_injection(
  mzroll_list,
  grouping_vars = "condition",
  peak_quant_var = "log2_abundance"
)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
\link{process_mzroll_multi}}

\item{grouping_vars}{character vector of sample variables to use for
grouping}

\item{peak_quant_var}{variable to plot in comparison}
}
\description{
Create a scatterplot of injection correlations.
}
\examples{
plot_compare_injection(
  nplug_mzroll_augmented,
  grouping_vars = "condition",
  peak_quant_var = "log2_abundance"
)
}
