% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{process_metabolon}
\alias{process_metabolon}
\title{Process Metabolon}
\usage{
process_metabolon(
  peak_areas,
  chem_anno,
  metadata_info,
  peak_floor_intensity = 4096
)
}
\arguments{
\item{peak_areas}{path to peak areas CSV file from metabolon results,
or table containing loaded peak areas}

\item{chem_anno}{path to chemical annotations CSV file from metabolon results,
or table containing loaded chemical annotations}

\item{metadata_info}{path to metadata CSV file from metabolon results,
or table containing loaded metadata}

\item{peak_floor_intensity}{value to use to replace NA peak measurements (missing values)}
}
\description{
Import results generated by metabolon into romic structure, using series of CSV
output files delivered by Metabolon as inputs.
}
\details{
An example dataset is available in the \code{MetabolonR} package,\cr
which may be installed via \code{remotes::install_github("miepstei/MetabolonR")}\cr

Once installed, execute the function via\cr\cr
\code{data <- process_metabolon(}\cr
\code{MetabolonR::metaboliteDataWide,}\cr
\code{MetabolonR::metaboliteInfo,}\cr
\code{MetabolonR::sampleInfo}\cr
\code{)}\cr
}
