% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{process_mzroll}
\alias{process_mzroll}
\title{Process mzRoll}
\usage{
process_mzroll(
  mzroll_db_path,
  only_identified = TRUE,
  validate = FALSE,
  method_tag = "",
  peakgroup_labels_to_keep = "*",
  peakgroup_labels_to_exclude = "",
  quant_col = "peakAreaTop"
)
}
\arguments{
\item{mzroll_db_path}{path to mzroll DB file}

\item{only_identified}{TRUE/FALSE, filter to only features which were
identified.}

\item{validate}{TRUE/FALSE, use meta-data to only name the subset of
features which were manually validated.}

\item{method_tag}{what method was run (for the purpose of matching
meta-data and aggregating).}

\item{peakgroup_labels_to_keep}{Peak groups containing any one of the
  characters in this string are retained.
\code{default = "*"} (retain all peak groups)}

\item{peakgroup_labels_to_exclude}{Peak groups containing any one of
the characters in this string are excluded.  Exclusion takes precedence
over inclusion. \code{default = ""} (do not exclude any peak groups).}

\item{quant_col}{name of quant column to use for standardization from the
  peaks table.
\code{default = "peakAreaTop"}}
}
\value{
a **triple_omic** from **romic** containing three tibbles:
\itemize{
  \item{features: one row per unique analyte (defined by a unique
    groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }
 And, a design list which tracks the variables in each table.
}
\description{
Process mzRoll
}
\examples{
process_mzroll(nplug_mzroll())
}
