% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{process_mzroll}
\alias{process_mzroll}
\title{Process mzRoll}
\usage{
process_mzroll(
  mzroll_db_path,
  standard_databases = NULL,
  sample_sheet_list = NULL,
  method_tag = "",
  only_identified = TRUE,
  validate = FALSE
)
}
\arguments{
\item{mzroll_db_path}{path to mzroll DB file}

\item{standard_databases}{connections to Calico standards and systematic
compound databases - generated with \link{configure_db_access}.}

\item{sample_sheet_list}{list of googlesheets information describing
experimental design - generated with \link{import_sample_sheet}.}

\item{method_tag}{what method was run (for the purpose of matching
meta-data and aggregating).}

\item{only_identified}{TRUE/FALSE, filter to only features which were
identified.}

\item{validate}{TRUE/FALSE, use meta-data to only name the subset of
features which were manually validated.}
}
\value{
an mzroll_list containing three tibbles:
\itemize{
  \item{peakgroups: one row per unique analyte (defined by a unique
    groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{peaks: one row per peak (samples x peakgroups)}
  }
}
\description{
Process mzRoll
}
\examples{
library(dplyr)

authutils::get_clamr_assets("X0083-M001A.mzrollDB") \%>\%
  process_mzroll(
    standard_databases = NULL,
    method_tag = "M001A",
    only_identified = TRUE,
    validate = FALSE
  )
}
