% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzroll.R
\name{process_mzroll_multi}
\alias{process_mzroll_multi}
\title{Process mzroll multi}
\usage{
process_mzroll_multi(
  mzroll_paths,
  standard_databases,
  sample_sheet_list,
  only_identified = TRUE,
  validate = FALSE
)
}
\arguments{
\item{mzroll_paths}{a tibble with two variables:
\itemize{
  \item{method_tag: a character vector to tag each dataset with},
  \item{mzroll_db_path: path to a mzrollDB file}
  }}

\item{standard_databases}{connections to Calico standards and systematic
compound databases - generated with \link{configure_db_access}.}

\item{sample_sheet_list}{list of googlesheets information describing
experimental design - generated with \link{import_sample_sheet}.}

\item{only_identified}{TRUE/FALSE, filter to only features which were
identified.}

\item{validate}{TRUE/FALSE, use meta-data to only name the subset of
features which were manually validated.}
}
\value{
an mzroll_list containing three tibbles:
\itemize{
  \item{peakgroups: one row per unique analyte (defined by a unique
    groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{peaks: one row per peak (samples x peakgroups)}
  }
}
\description{
Aggregate multiple mzrollDB datasets which possess the same sample metadata
}
