% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_mzroll_list}
\alias{test_mzroll_list}
\title{Test MzRoll List}
\usage{
test_mzroll_list(mzroll_list, fast_check = TRUE)
}
\arguments{
\item{mzroll_list}{output of \link{process_mzroll} or
\link{process_mzroll_multi}}

\item{fast_check}{if `TRUE` then skip some checks which are slow and that are
generally only needed when a `tomic` object is first created

\itemize{
  \item{features: one row per unique analyte (defined by a
    unique groupId)},
  \item{samples: one row per unique sample (defined by a unique sampleId)},
  \item{measurements: one row per peak (samples x peakgroups)}
  }}
}
\description{
Test MzRoll List
}
